Move EVM memory to a flat bytes array
Performance measurements have indicated that the list of Bytes32 for
memory is not performant enough. Moving to a byte array reduces wrapper
object thrashing as well as simplifying cross-word boundary logic (there
are no boundaries).

+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/vm/MessageFrame.java
@@ -553,7 +553,7 @@ public class MessageFrame {
    * @param offset The offset in memory
    * @param length The length of the memory access
    */
-  public void expandMemory(final long offset, final int length) {
+  public void expandMemory(final int offset, final int length) {
     memory.ensureCapacityForBytes(offset, length);
   }
